Moved SurveyQuestion component to src root
